IQ-15 #comment changed constructor from using a float to a double

+++ src/main/java/com/qa/ims/persistence/dao/ProductDAO.java
@@ -86,7 +86,7 @@ public class ProductDAO implements Dao<Product> {
 				PreparedStatement statement = connection
 						.prepareStatement("INSERT INTO products (product_name, product_value) VALUES (?, ?)");) {
 			statement.setString(1, product.getProduct_name());
-			statement.setFloat(2, product.getProduct_value());
+			statement.setDouble(2, product.getProduct_value());
 			statement.executeUpdate();
 			return readLatest();
 		} catch (Exception e) {
@@ -110,7 +110,7 @@ public class ProductDAO implements Dao<Product> {
 				PreparedStatement statement = connection
 						.prepareStatement("UPDATE products SET product_name = ?, product_value = ? WHERE id = ?");) {
 			statement.setString(1, product.getProduct_name());
-			statement.setFloat(2, product.getProduct_value());
+			statement.setDouble(2, product.getProduct_value());
 			statement.executeUpdate();
 			return read(product.getProduct_id());
 		} catch (Exception e) {
@@ -142,7 +142,7 @@ public class ProductDAO implements Dao<Product> {
 	public Product modelFromResultSet(ResultSet resultSet) throws SQLException {
 		Long product_id = resultSet.getLong("product_id");
 		String product_name = resultSet.getString("product_name");
-		float product_value = resultSet.getFloat("product_value");
+		Double product_value = resultSet.getDouble("product_value");
 		return new Product(product_id, product_name, product_value);
 	}
